refactor: sprint 3 — pytest infrastructure + split api/actionable.py into a package
Tests (audit roadmap: foundation for all future refactoring):
- New tests/ directory at repo root with first real unit-test suite.
- tests/conftest.py: env-var bootstrap before any libinv import; stubs
  for SCIO models (DiscoveredPackage, VulnerablePath, ScanpipeProject);
  reusable flask_app_client and reset_daemon_shutdown_flag fixtures.
- tests/test_helpers.py: explode_git_url (4 cases incl. unsupported
  scheme raising ValueError), retry_on_exception (count + re-raise),
  send_to_slack (timeout=10 + RequestException is logged not raised).
- tests/test_issue_reporter.py: empty / only-P0 / only-Other / mixed
  rendering, commit_id and jenkins_url append behavior, magnifier emoji
  for empty suggested_versions, table header invariant.
- tests/test_auth.py: 6 cases via flask test client — GET unaffected,
  PUT/POST/PATCH/DELETE rejected without/wrong/correct token, 503 when
  LIBINV_API_TOKEN unset.
- tests/test_semgrep_runner.py: 5 cases verifying argv-list call (not
  shell), shell=False, untrusted strings pass through unchanged,
  timeout=3600.
- tests/test_daemon_shutdown.py: SIGTERM/SIGINT handler flips
  _shutdown_requested; restored via fixture.
- tests/test_session_scope.py: commit on clean exit, rollback on
  exception, remove() always called.
- pytest.ini: testpaths = tests libinv; --doctest-modules picks up the
  4 explode_git_url doctests automatically; ignore globs for the
  reflection-time-DB modules (scio_models, scancodeio) and alembic.
- Makefile tests target: single `pytest` invocation under coverage (was
  the old doctests-only chain).
- 38 tests pass (37 unit + 1 doctest) in ~0.3s.

api/actionable.py split (audit #17):
- The 1218-LOC god-route is gone. Replaced by libinv/api/actionable/
  package:
    __init__.py         — Blueprint declaration + re-exports
                          fetch_repository + imports view modules to
                          register routes.
    _common.py          — fetch_repository helper (lifted verbatim).
    dashboards.py       — /v2/ and /v3/ (preserves the v2/v3 split for
                          this sprint; consolidation deferred to Sprint
                          4 where tests catch regressions).
    package_details.py  — /v3/package-details
    repositories.py     — /v3/repositories
    statistics.py       — /v3/statistics (still ~415 LOC; CTE
                          consolidation of its 14 serial queries is
                          Sprint 4 work).
    package_scan.py     — /v3/package_scan GET +
                          /v3/request_package_scan POST.
- template_folder dropped from the Blueprint — Flask app's global
  template_folder=templates covers it; the otherwise-relative path
  would resolve to a non-existent libinv/api/actionable/templates/.
- Public-API preserved: libinv/api/app.py:8 still imports `actionable`
  unchanged; libinv/api/compare_builds.py:13 still imports
  `fetch_repository` unchanged (re-exported from the package
  __init__.py).
- All Sprint 0-2 fixes inside the routes preserved verbatim (list-comp
  safe_version rename in dashboards.py, indented statistics_dashboard
  with block, bare-except → logger, corrected status codes).
- Unused `import re` from the original file dropped (audit confirmed).

conn deprecation (audit prep for Sprint 4):
- libinv/base.py: conn (the ScopedSession alias) gets a deprecation
  comment pointing callers at session_scope() / explicit session
  parameters. Kept working for now; the ~30 existing call sites are
  migrated incrementally in later sprints.

Bug fixes flagged by the test agent and fixed in this sprint:
- libinv/conftest.py was missing a DiscoveredPackage MagicMock on the
  fake scio_models module, breaking doctest collection of
  libinv/cli/epss.py — added.

Verification:
- Full libinv tree + tests/ + alembic/ pass python3 -m py_compile.
- `pytest` runs green: 38 passed in ~0.3s.
