+++ libinv/conftest.py
@@ -8,3 +8,4 @@ sys.modules["libinv.scio_models"] = fake_scio_models
 # list all classes that reflect SCIO related tables here
 fake_scio_models.VulnerablePath = MagicMock()
 fake_scio_models.ScanpipeProject = MagicMock()
+fake_scio_models.DiscoveredPackage = MagicMock()
