+++ pytest.ini
@@ -1,0 +1,12 @@
+[pytest]
+testpaths = tests libinv
+addopts =
+    -ra
+    --strict-markers
+    --strict-config
+    --doctest-modules
+    --ignore-glob=*scio_models.py
+    --ignore-glob=*scancodeio.py
+    --ignore=alembic
+filterwarnings =
+    ignore::DeprecationWarning:sqlalchemy.*
